Circular slider

+++ lib/components/time_slider/time_slider.dart
@@ -4,7 +4,7 @@ import 'package:sleep_timer/utils/app_variables.dart';
 
 class TimeSlider extends StatelessWidget {
   late bool isStart;
-  late int timerValue;
+  late double timerValue;
   late Function onChange;
 
   TimeSlider({
@@ -18,7 +18,7 @@ class TimeSlider extends StatelessWidget {
   Widget build(BuildContext context) {
     double screenWidth = MediaQuery.of(context).size.width;
     double sliderWidth = screenWidth - AppVariables.MAIN_PADDING * 2;
-    int maxTime = AppVariables.MAX_TIME;
+    double maxTime = AppVariables.MAX_TIME;
 
     return Container(
       clipBehavior: Clip.hardEdge,
